:truck: api: move getShareableCertificate method from lib to src
